토큰 적용중

+++ back/src/main/java/com/zooflix/be_zooflix/domain/stockSubscribe/service/StockSubscribeService.java
@@ -65,8 +65,11 @@ public class StockSubscribeService {
      * 3.2 주식 정기 구독 해지
      *
      */
-    public int terminationSubscribe(int stockSubscribeNo){
+    public int terminationSubscribe(int stockSubscribeNo, int userNo){
         StockSubscribe subscribe = stockSubscribeRepository.findByStockSubscribeNo(stockSubscribeNo);
+        if (userNo != subscribe.getUser().getUserNo()) {
+            throw new RuntimeException("토큰 정보와 일치하는 유저가 아닙니다!");
+        }
         stockSubscribeRepository.delete(subscribe);
 
         //예약주문된 내역이 있다면 주문취소하기
